requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
redis>=4.2
google-generativeai
psycopg[binary]
sqlalchemy
alembic 
asyncpg
psycopg2
aiogram
